+++ src/formatters/network.js
@@ -1,7 +1,8 @@
 /* eslint-disable no-underscore-dangle */
-const { includes } = require('lodash');
+const { includes, groupBy } = require('lodash');
 const {
   entityPrimaryKeyProperty,
+  entityAttributesProperty,
   sessionProperty,
   egoProperty,
   nodeExportIDProperty,
@@ -10,22 +11,29 @@ const {
   ncTargetUUID,
   edgeSourceProperty,
   edgeTargetProperty,
+  caseProperty,
+  sessionFinishTimeProperty,
+  sessionStartTimeProperty,
+  sessionExportTimeProperty,
+  protocolProperty,
+  protocolName,
+  codebookHashProperty
 } = require('@codaco/shared-consts');
 const { getAttributePropertyFromCodebook } = require('./graphml/helpers');
-const { getEntityAttributes } = require('../utils/general');
+const general = require('../utils/general');
 
 // Determine which variables to include
 // TODO: Move this to CSV formatter, since only CSV uses it
 const processEntityVariables = (entity, entityType, codebook, exportSettings) => ({
   ...entity,
-  attributes: Object.keys(getEntityAttributes(entity)).reduce(
+  attributes: Object.keys(general.getEntityAttributes(entity)).reduce(
     (
       accumulatedAttributes,
       attributeUUID,
     ) => {
       const attributeName = getAttributePropertyFromCodebook(codebook, entityType, entity, attributeUUID, 'name');
       const attributeType = getAttributePropertyFromCodebook(codebook, entityType, entity, attributeUUID, 'type');
-      const attributeData = getEntityAttributes(entity)[attributeUUID];
+      const attributeData = general.getEntityAttributes(entity)[attributeUUID];
 
       if (attributeType === 'categorical') {
         const attributeOptions = getAttributePropertyFromCodebook(codebook, entityType, entity, attributeUUID, 'options') || [];
@@ -225,11 +233,118 @@ const unionOfNetworks = (sessionsByProtocol) => Object.keys(sessionsByProtocol)
     };
   }, {});
 
+// Function designed to mirror the flow in FileExportManager.exportSessions()
+const processMockNetworks = (networkCollection, unify) => {
+  const sessionsWithEgo = insertEgoIntoSessionNetworks(networkCollection);
+  const sessionsWithResequencedIDs = resequenceIds(sessionsWithEgo);
+  const sessionsByProtocol = groupBy(sessionsWithResequencedIDs, `sessionVariables.${protocolProperty}`);
+
+  if (!unify) {
+    return sessionsByProtocol;
+  }
+  return unionOfNetworks(sessionsByProtocol);
+};
+
+
+const mockNetwork = {
+  nodes: [
+    { [entityPrimaryKeyProperty]: '1', type: 'mock-node-type', [entityAttributesProperty]: { 'mock-uuid-1': 'Dee', 'mock-uuid-2': 40, 'mock-uuid-3': { x: 0, y: 0 }, 'mock-uuid-4': true, 'mock-uuid-5': null } },
+    { [entityPrimaryKeyProperty]: '2', type: 'mock-node-type', [entityAttributesProperty]: { 'mock-uuid-1': 'Carl', 'mock-uuid-2': 0, 'mock-uuid-3': { x: 0, y: 0 }, 'mock-uuid-4': false, 'mock-uuid-5': null } },
+    { [entityPrimaryKeyProperty]: '3', type: 'mock-node-type', [entityAttributesProperty]: { 'mock-uuid-1': 'Jumbo', 'mock-uuid-2': 50, 'mock-uuid-3': null, 'mock-uuid-4': true, 'mock-uuid-5': null } },
+    { [entityPrimaryKeyProperty]: '4', type: 'mock-node-type', [entityAttributesProperty]: { 'mock-uuid-1': 'Francis', 'mock-uuid-2': 10, 'mock-uuid-3': { x: 0, y: 0 }, 'mock-uuid-4': null, 'mock-uuid-5': null } },
+  ],
+  edges: [
+    { from: '1', to: '2', type: 'mock-edge-type' },
+  ],
+  ego: {
+    [entityPrimaryKeyProperty]: 'ego-id-1',
+    [entityAttributesProperty]: {
+      'mock-uuid-1': 'Dee',
+      'mock-uuid-2': 40,
+      'mock-uuid-3': false,
+    },
+  },
+  sessionVariables: {
+    [caseProperty]: 123,
+    [protocolName]: 'protocol name',
+    [protocolProperty]: 'protocol-uid-1',
+    [sessionProperty]: 'session-id-1',
+    [sessionStartTimeProperty]: 100,
+    [sessionFinishTimeProperty]: 200,
+    [sessionExportTimeProperty]: 300,
+    [codebookHashProperty]: '14fa461bf4b98155e82adc86532938553b4d33a9',
+  },
+};
+
+const mockNetwork2 = {
+  nodes: [
+    { [entityPrimaryKeyProperty]: '10', type: 'mock-node-type', [entityAttributesProperty]: { 'mock-uuid-1': 'Jimbo', 'mock-uuid-2': 20, 'mock-uuid-3': { x: 10, y: 50 } } },
+    { [entityPrimaryKeyProperty]: '20', type: 'mock-node-type', [entityAttributesProperty]: { 'mock-uuid-1': 'Jambo', 'mock-uuid-2': 30, 'mock-uuid-3': { x: 20, y: 20 } } },
+  ],
+  edges: [
+    { from: '10', to: '20', type: 'mock-edge-type' },
+  ],
+  ego: {
+    [entityPrimaryKeyProperty]: 'ego-id-10',
+    [entityAttributesProperty]: {
+      'mock-uuid-1': 'Dee',
+      'mock-uuid-2': 40,
+      'mock-uuid-3': true,
+    },
+  },
+  sessionVariables: {
+    [caseProperty]: 456,
+    [protocolName]: 'protocol name',
+    [protocolProperty]: 'protocol-uid-1',
+    [sessionProperty]: 'session-id-2',
+    [sessionStartTimeProperty]: 1000,
+    [sessionFinishTimeProperty]: 2000,
+    [sessionExportTimeProperty]: 3000,
+    [codebookHashProperty]: '14fa461bf4b98155e82adc86532938553b4d33a9',
+  },
+};
+
+const mockCodebook = {
+  ego: {
+    variables: {
+      'mock-uuid-1': { name: 'egoName', type: 'string' },
+      'mock-uuid-2': { name: 'egoAge', type: 'number' },
+      'mock-uuid-3': { name: 'boolVar', type: 'boolean' },
+    },
+  },
+  node: {
+    'mock-node-type': {
+      name: 'person',
+      variables: {
+        'mock-uuid-1': { name: 'firstName', type: 'string' },
+        'mock-uuid-2': { name: 'age', type: 'number' },
+        'mock-uuid-3': { name: 'layout', type: 'layout' },
+        'mock-uuid-4': { name: 'boolWithValues', type: 'boolean' },
+        'mock-uuid-5': { name: 'nullBool', type: 'boolean' },
+        'mock-uuid-6': { name: 'unusedBool', type: 'boolean' },
+      },
+    },
+  },
+  edge: {
+    'mock-edge-type': {
+      name: 'peer',
+    },
+    'mock-edge-type-2': {
+      name: 'likes',
+    },
+  },
+};
+
+
 module.exports = {
+  mockCodebook,
+  mockNetwork,
+  mockNetwork2,
   processEntityVariables,
   insertNetworkEgo,
   insertEgoIntoSessionNetworks,
   partitionNetworkByType,
   resequenceIds,
   unionOfNetworks,
+  processMockNetworks,
 };
